requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
coverage==6.3.1
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
Faker==12.0.1
iniconfig==1.1.1
Markdown==3.3.6
nose2==0.10.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
